Refactor
- Functions instead of injection
- Additional comments
- Remove dependency between feature:login and okhttp
