+++ app/src/test/java/app/pachli/di/FakeNetworkModule.kt
@@ -29,7 +29,6 @@ import dagger.hilt.testing.TestInstallIn
 import java.util.Date
 import javax.inject.Singleton
 import okhttp3.OkHttpClient
-import okhttp3.tls.HandshakeCertificates
 import org.mockito.kotlin.mock
 
 @TestInstallIn(
@@ -48,10 +47,6 @@ object FakeNetworkModule {
     @Provides
     @Singleton
     fun providesHttpClient(): OkHttpClient = mock()
-
-    @Provides
-    @Singleton
-    fun providesHandshakeCertificats(): HandshakeCertificates = mock()
 
     @Provides
     @Singleton
